VideoConvertor/MediaProcessor.cs
using System.Resources;
using Xabe.FFmpeg;

namespace VideoConvertor
{
    public class MediaProcessor
    {
        // private readonly string _mpegExePath;

        // سازنده کلاس که مسیر mpeg.exe را دریافت می‌کند
        public MediaProcessor()
        {
            // _mpegExePath = "Tools\\ffmpeg\\bin\\ffmpeg.exe";
        }

        // متدی برای دریافت متادیتای فیلم
        public async Task<IMediaInfo> GetVideoMetadata(string inputFilePath)
        {
            IMediaInfo mediaInfo = await FFmpeg.GetMediaInfo(inputFilePath);
            return mediaInfo;
        }

        // متدی برای گرفتن اسکرین شات از فریم فیلم
        public async Task TakeScreenshot(string inputFilePath, TimeSpan timestamp, string outputImagePath)
        {
            IConversion conversion = await FFmpeg.Conversions.FromSnippet.Snapshot(inputFilePath, outputImagePath, timestamp);
            IConversionResult result = await conversion.Start();
        }

        //// متدی برای برش ویدئو
        //public void TrimVideo(string inputFilePath, TimeSpan startTime, TimeSpan endTime, string outputFilePath)
        //{
        //    FFMpeg.SubVideo(inputFilePath, outputFilePath, startTime, endTime);
        //}

        //// متدی برای استخراج صوت از ویدئو
        //public void ExtractAudio(string inputFilePath, string outputAudioPath)
        //{
        //    FFMpeg.ExtractAudio(inputFilePath, outputAudioPath);
        //}

        // متدی برای تولید صوت از متن
        public void GenerateAudioFromText(string text, string outputAudioPath)
        {
            // این کار ممکن است نیاز به ابزار دیگری غیر از mpeg.exe داشته باشد.
            // فرض می‌کنیم یک ابزار Text-to-Speech دارید که می‌توان از آن استفاده کرد.
            // در اینجا تنها شبیه‌سازی کرده‌ایم.

            throw new NotImplementedException("Text-to-Speech functionality needs an external tool.");
        }

    }
}
